Experimental Changes
1. Parallel Shader Compilation
2. 16 Bit Pointers
3. Inlined Chunk Scheduling

+++ src/main/java/me/jellysquid/mods/sodium/client/gl/attribute/GlVertexFormat.java
@@ -62,7 +62,7 @@ public class GlVertexFormat<T extends Enum<T>> implements BufferVertexFormat {
             this.stride = stride;
         }
 
-        public Builder<T> addElement(T type, int pointer, GlVertexAttributeFormat format, int count, boolean normalized) {
+        public Builder<T> addElement(T type, short pointer, GlVertexAttributeFormat format, int count, boolean normalized) {
             return this.addElement(type, new GlVertexAttribute(format, count, normalized, pointer, this.stride));
         }
